refactor: move relevant util code into `controller` package
Also remove dead code.
