+++ internal/controller/add_core_set.go
@@ -9,6 +9,7 @@ import (
 	"github.com/cisco-open/k8s-objectmatcher/patch"
 	appsv2beta1 "github.com/emqx/emqx-operator/api/v2beta1"
 	config "github.com/emqx/emqx-operator/internal/controller/config"
+	util "github.com/emqx/emqx-operator/internal/controller/util"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	k8sErrors "k8s.io/apimachinery/pkg/api/errors"
@@ -134,9 +135,9 @@ func getNewStatefulSet(instance *appsv2beta1.EMQX, conf *config.Conf) *appsv1.St
 	preSts.Labels = appsv2beta1.CloneAndAddLabel(preSts.Labels, appsv2beta1.LabelsPodTemplateHashKey, podTemplateSpecHash)
 	preSts.Spec.Selector = appsv2beta1.CloneSelectorAndAddLabel(preSts.Spec.Selector, appsv2beta1.LabelsPodTemplateHashKey, podTemplateSpecHash)
 	preSts.Spec.Template.Labels = appsv2beta1.CloneAndAddLabel(preSts.Spec.Template.Labels, appsv2beta1.LabelsPodTemplateHashKey, podTemplateSpecHash)
-	preSts.Spec.Template.Spec.Containers[0].Ports = appsv2beta1.MergeContainerPorts(
+	preSts.Spec.Template.Spec.Containers[0].Ports = util.MergeContainerPorts(
 		preSts.Spec.Template.Spec.Containers[0].Ports,
-		appsv2beta1.TransServicePortsToContainerPorts(svcPorts),
+		util.MapServicePortsToContainerPorts(svcPorts),
 	)
 	for _, p := range preSts.Spec.Template.Spec.Containers[0].Ports {
 		if p.Name == "dashboard" {
