+++ internal/controller/load_state.go
@@ -8,6 +8,7 @@ import (
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
 	k8s "sigs.k8s.io/controller-runtime/pkg/client"
 )
 
@@ -30,6 +31,16 @@ func (r *reconcileState) podsWithRole(role string) []*corev1.Pod {
 	var list []*corev1.Pod
 	for _, pod := range r.pods {
 		if pod.Labels[appsv2beta1.LabelsDBRoleKey] == role {
+			list = append(list, pod)
+		}
+	}
+	return list
+}
+
+func (r *reconcileState) podsManagedBy(uid types.UID) []*corev1.Pod {
+	var list []*corev1.Pod
+	for _, pod := range r.pods {
+		if metav1.GetControllerOf(pod) != nil && metav1.GetControllerOf(pod).UID == uid {
 			list = append(list, pod)
 		}
 	}
