+++ internal/controller/update_pod_conditions.go
@@ -3,6 +3,7 @@ package controller
 import (
 	emperror "emperror.dev/errors"
 	appsv2beta1 "github.com/emqx/emqx-operator/api/v2beta1"
+	util "github.com/emqx/emqx-operator/internal/controller/util"
 	corev1 "k8s.io/api/core/v1"
 )
 
@@ -12,7 +13,7 @@ type updatePodConditions struct {
 
 func (u *updatePodConditions) reconcile(r *reconcileRound, instance *appsv2beta1.EMQX) subResult {
 	for _, pod := range r.state.pods {
-		onServingCondition := appsv2beta1.FindPodCondition(pod, appsv2beta1.PodOnServing)
+		onServingCondition := util.FindPodCondition(pod, appsv2beta1.PodOnServing)
 		if onServingCondition == nil {
 			onServingCondition = &corev1.PodCondition{
 				Type:   appsv2beta1.PodOnServing,
@@ -21,24 +22,24 @@ func (u *updatePodConditions) reconcile(r *reconcileRound, instance *appsv2beta1
 		}
 
 		if r.state.partOfUpdateSet(pod, instance) {
-			cond := appsv2beta1.FindPodCondition(pod, corev1.ContainersReady)
+			cond := util.FindPodCondition(pod, corev1.ContainersReady)
 			if cond != nil && cond.Status == corev1.ConditionTrue {
 				status := u.checkRebalanceStatus(r, pod)
-				appsv2beta1.ChangeCondition(onServingCondition, status)
+				util.SwitchPodConditionStatus(onServingCondition, status)
 			}
 		} else {
 			if r.state.partOfCurrentSet(pod, instance) {
 				// When available condition is true, need clean currentSts / currentRs pod
 				if instance.Status.IsConditionTrue(appsv2beta1.Available) {
-					cond := appsv2beta1.FindPodCondition(pod, corev1.ContainersReady)
+					cond := util.FindPodCondition(pod, corev1.ContainersReady)
 					if cond != nil && cond.Status == corev1.ConditionTrue {
-						appsv2beta1.ChangeCondition(onServingCondition, corev1.ConditionFalse)
+						util.SwitchPodConditionStatus(onServingCondition, corev1.ConditionFalse)
 					}
 				}
 			}
 		}
 
-		err := updatePodCondition(r.ctx, u.Client, pod, *onServingCondition)
+		err := util.UpdatePodCondition(r.ctx, u.Client, pod, *onServingCondition)
 		if err != nil {
 			return subResult{err: emperror.Wrapf(err, "failed to update pod %s status", pod.Name)}
 		}
